\name{mRNA_matrix}
\alias{mRNA_matrix}
\docType{data}
\title{
The expression data 
}
\description{
An example of breast cancer expression data with 2000 genes and 100 samples. 
}
\usage{data("mRNA_matrix")}
\format{
  The format is:
 num [1:2000, 1:100] 8.97 6.04 6.51 5.46 5.52 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:2000] "780" "5982" "3310" "7849" ...
  ..$ : chr [1:100] "GSM79114" "GSM79115" "GSM79116" "GSM79118" ...
}
\examples{
data(mRNA_matrix)
}
\keyword{datasets}
