\name{pathSet}
\alias{pathSet}
\docType{data}
\title{
Pathway set
}
\description{
	\code{pathSet} is composed of 306 KEGG pathways. Each pathway is represented as a vector of pathway member genes and metabolites.
}
\usage{data("pathSet")}
\format{
   A list of 306 KEGG pathways.
}
\examples{
data(pathSet)
}
\keyword{datasets}
